+++ _fused_quant.py
@@ -22,7 +22,7 @@ def _quant_kernel(
     valid = rows < T
 
     x = tl.load(
-        x_ptr + batch * s_b + rows[:, None] * s_t + head * s_h + d[None, :],
+        x_ptr + batch * s_b + rows[:, None].to(tl.int64) * s_t + head * s_h + d[None, :],
         mask=valid[:, None], other=0.0,
     ).to(tl.float32)
 
@@ -35,7 +35,7 @@ def _quant_kernel(
     xi = tl.extra.cuda.libdevice.round(x / s_safe[:, None])
     xi = tl.minimum(tl.maximum(xi, -127.0), 127.0).to(tl.int8)
 
-    offs_out = ((batch * T + rows[:, None]) * H + head) * D + d[None, :]
+    offs_out = ((batch * T + rows[:, None]).to(tl.int64) * H + head) * D + d[None, :]
     tl.store(xi_ptr + offs_out, xi, mask=valid[:, None])
     tl.store(xs_ptr + (batch * T + rows) * H + head, s_safe, mask=valid)
